warp::filters

Module body

Body filters

Filters that extract a body for a route.

Structs

BodyConsumedMultipleTimes 🔒

BodyDeserializeError

An error used in rejections when deserializing a request body fails.

BodyReadError 🔒

Form 🔒

Json 🔒

Traits

Decode 🔒

Functions

aggregate

Returns a Filter that matches any request and extracts a Future of an aggregated body.

body 🔒

bytes

Returns a Filter that matches any request and extracts a Future of a concatenated body.

content_length_limit

Require a content-length header to have a value no greater than some limit.

form

Returns a Filter that matches any request and extracts a Future of a form encoded body.

is_content_type 🔒

json

Returns a Filter that matches any request and extracts a Future of a JSON-decoded body.

stream

Create a Filter that extracts the request body as a futures::Stream.

Type Aliases

BoxError 🔒