warp::filters::body
Function stream

pub fn stream() -> impl Filter<Extract = (impl Stream<Item = Result<impl Buf, Error>>,), Error = Rejection> + Copy

Create a Filter that extracts the request body as a futures::Stream.

If other filters have already extracted the body, this filter will reject with a 500 Internal Server Error.

For example usage, please take a look at examples/stream.rs.

Warning

This does not have a default size limit, it would be wise to use one to prevent a overly large request from using too much memory.