Struct warp::filters::cors::Builder

pub struct Builder { /* private fields */ }

A constructed via warp::cors().

Implementations

impl Builder

pub fn allow_credentials(self, allow: bool) -> Self

Sets whether to add the Access-Control-Allow-Credentials header.

pub fn allow_method<M>(self, method: M) -> Self where
Method: TryFrom<M>,

Adds a method to the existing list of allowed request methods.

Panics

Panics if the provided argument is not a valid http::Method.

pub fn allow_methods<I>(self, methods: I) -> Self where
I: IntoIterator,
Method: TryFrom<I::Item>,

Adds multiple methods to the existing list of allowed request methods.

Panics

Panics if the provided argument is not a valid http::Method.

pub fn allow_header<H>(self, header: H) -> Self where
HeaderName: TryFrom<H>,

Adds a header to the list of allowed request headers.

Note: These should match the values the browser sends via Access-Control-Request-Headers, e.g. content-type.

Panics

Panics if the provided argument is not a valid http::header::HeaderName.

pub fn allow_headers<I>(self, headers: I) -> Self where
I: IntoIterator,
HeaderName: TryFrom<I::Item>,

Adds multiple headers to the list of allowed request headers.

Note: These should match the values the browser sends via Access-Control-Request-Headers, e.g.content-type.

Panics

Panics if any of the headers are not a valid http::header::HeaderName.

pub fn expose_header<H>(self, header: H) -> Self where
HeaderName: TryFrom<H>,

Adds a header to the list of exposed headers.

Panics

Panics if the provided argument is not a valid http::header::HeaderName.

pub fn expose_headers<I>(self, headers: I) -> Self where
I: IntoIterator,
HeaderName: TryFrom<I::Item>,

Adds multiple headers to the list of exposed headers.

Panics

Panics if any of the headers are not a valid http::header::HeaderName.

pub fn allow_any_origin(self) -> Self

Sets that any Origin header is allowed.

Warning

This can allow websites you didn’t intend to access this resource, it is usually better to set an explicit list.

pub fn allow_origin(self, origin: impl IntoOrigin) -> Self

Add an origin to the existing list of allowed Origins.

Panics

Panics if the provided argument is not a valid Origin.

pub fn allow_origins<I>(self, origins: I) -> Self where
I: IntoIterator,
I::Item: IntoOrigin,

Add multiple origins to the existing list of allowed Origins.

Panics

Panics if the provided argument is not a valid Origin.

pub fn max_age(self, seconds: impl Seconds) -> Self

Sets the Access-Control-Max-Age header.

Example

use std::time::Duration;
use warp::Filter;

let cors = warp::cors()
    .max_age(30) // 30u32 seconds
    .max_age(Duration::from_secs(30)); // or a Duration

pub fn build(self) -> Cors

Builds the Cors wrapper from the configured settings.

This step isn’t required, as the Builder itself can be passed to Filter::with. This just allows constructing once, thus not needing to pay the cost of “building” every time.