Struct warp::filters::cors::Cors

pub struct Cors { /* fields omitted */ }

A wrapping filter constructed via warp::cors().

Implementations

impl Cors

pub fn allow_credentials(self, allow: bool) -> Self

Sets whether to add the Access-Control-Allow-Credentials header.

pub fn allow_method<M>(self, method: M) -> Self where
Method: HttpTryFrom<M>,

Adds a method to the existing list of allowed request methods.

Panics

Panics if the provided argument is not a valid http::Method.

pub fn allow_methods<I>(self, methods: I) -> Self where
I: IntoIterator,
Method: HttpTryFrom<I::Item>,

Adds multiple methods to the existing list of allowed request methods.

Panics

Panics if the provided argument is not a valid http::Method.

pub fn allow_header<H>(self, header: H) -> Self where
HeaderName: HttpTryFrom<H>,

Adds a header to the list of allowed request headers.

Panics

Panics if the provided argument is not a valid http::header::HeaderName.

pub fn allow_headers<I>(self, headers: I) -> Self where
I: IntoIterator,
HeaderName: HttpTryFrom<I::Item>,

Adds multiple headers to the list of allowed request headers.

Panics

Panics if any of the headers are not a valid http::header::HeaderName.

pub fn expose_header<H>(self, header: H) -> Self where
HeaderName: HttpTryFrom<H>,

Adds a header to the list of exposed headers.

Panics

Panics if the provided argument is not a valid http::header::HeaderName.

pub fn expose_headers<I>(self, headers: I) -> Self where
I: IntoIterator,
HeaderName: HttpTryFrom<I::Item>,

Adds multiple headers to the list of exposed headers.

Panics

Panics if any of the headers are not a valid http::header::HeaderName.

pub fn allow_any_origin(self) -> Self

Sets that any Origin header is allowed.

Warning

This can allow websites you didn’t instead to access this resource, it is usually better to set an explicit list.

pub fn allow_origin(self, origin: impl IntoOrigin) -> Self

Add an origin to the existing list of allowed Origins.

Panics

Panics if the provided argument is not a valid Origin.

pub fn allow_origins<I>(self, origins: I) -> Self where
I: IntoIterator,
I::Item: IntoOrigin,

Add multiple origins to the existing list of allowed Origins.

Panics

Panics if the provided argument is not a valid Origin.

pub fn max_age(self, seconds: impl Seconds) -> Self

Sets the Access-Control-Max-Age header.

Example

use std::time::Duration;
use warp::Filter;

let cors = warp::cors()
    .max_age(30) // 30u32 seconds
    .max_age(Duration::from_secs(30)); // or a Duration