Module warp::filters::method

HTTP Method filters.

The filters deal with the HTTP Method part of a request. Several here will match the request Method, and if not matched, will reject the request with a 405 Method Not Allowed.

There is also warp::method(), which never rejects a request, and just extracts the method to be used in your filter chains.

Functions

delete
Create a Filter that requires the request method to be DELETE.

get
Create a Filter that requires the request method to be GET.

head
Create a Filter that requires the request method to be HEAD.

method
Extract the Method from the request.

options
Create a Filter that requires the request method to be OPTIONS.

patch
Create a Filter that requires the request method to be PATCH.

post
Create a Filter that requires the request method to be POST.

put
Create a Filter that requires the request method to be PUT.