warp::filters::multipart

Function form

pub fn form() -> FormOptions

Create a Filter to extract a multipart/form-data body from a request.

The extracted FormData type is a Stream of Parts, and each Part in turn is a Stream of bytes.