Function warp::filters::sse::keep

pub fn keep<S>(
    event_stream: S, 
    keep_interval: impl Into<Option<Duration>>
) -> impl Stream<Item = impl ServerSentEvent + Send + 'static, Error = impl StdError + Send + Sync + 'static> + Send + 'static where
    S: Stream + Send + 'static,
    S::Item: ServerSentEvent + Send,
    S::Error: StdError + Send + Sync + 'static, 

Keeps event source connection when no events sent over a some time.

Some proxy servers may drop HTTP connection after a some timeout of inactivity. This function helps to prevent such behavior by sending dummy events with single empty comment field (i.e. ":" only) each keep_interval of inactivity.

By default this time interval between events equals to 15 seconds.

See notes.