Struct warp::filters::sse::Event

pub struct Event { /* fields omitted */ }

Server-sent event

Implementations

impl Event

pub fn data<T: Into<String>>(self, data: T) -> Event

Set Server-sent event data data field(s) (“data:”)

Examples found in repository

examples/sse.rs

≺ ≻↕

9
10
11

fn sse_counter(counter: u64) -> Result<Event, Infallible> {
    Ok(warp::sse::Event::default().data(counter.to_string()))
}

examples/sse_chat.rs

≺ ≻↕

78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101

fn user_connected(users: Users) -> impl Stream<Item = Result<Event, warp::Error>> + Send + 'static {
    // Use a counter to assign a new unique ID for this user.
    let my_id = NEXT_USER_ID.fetch_add(1, Ordering::Relaxed);

    eprintln!("new chat user: {}", my_id);

    // Use an unbounded channel to handle buffering and flushing of messages
    // to the event source...
    let (tx, rx) = mpsc::unbounded_channel();
    let rx = UnboundedReceiverStream::new(rx);

    tx.send(Message::UserId(my_id))
        // rx is right above, so this cannot fail
        .unwrap();

    // Save the sender in our list of connected users.
    users.lock().unwrap().insert(my_id, tx);

    // Convert messages into Server-Sent Events and return resulting stream.
    rx.map(|msg| match msg {
        Message::UserId(my_id) => Ok(Event::default().event("user").data(my_id.to_string())),
        Message::Reply(reply) => Ok(Event::default().data(reply)),
    })
}

pub fn json_data<T: Serialize>(self, data: T) -> Result<Event, Error>

Set Server-sent event data data field(s) (“data:”)

pub fn comment<T: Into<String>>(self, comment: T) -> Event

Set Server-sent event comment Comment field (“:”)

pub fn event<T: Into<String>>(self, event: T) -> Event

Set Server-sent event event Event name field (“event:”)

Examples found in repository

examples/sse_chat.rs

≺ ≻↕

78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101

fn user_connected(users: Users) -> impl Stream<Item = Result<Event, warp::Error>> + Send + 'static {
    // Use a counter to assign a new unique ID for this user.
    let my_id = NEXT_USER_ID.fetch_add(1, Ordering::Relaxed);

    eprintln!("new chat user: {}", my_id);

    // Use an unbounded channel to handle buffering and flushing of messages
    // to the event source...
    let (tx, rx) = mpsc::unbounded_channel();
    let rx = UnboundedReceiverStream::new(rx);

    tx.send(Message::UserId(my_id))
        // rx is right above, so this cannot fail
        .unwrap();

    // Save the sender in our list of connected users.
    users.lock().unwrap().insert(my_id, tx);

    // Convert messages into Server-Sent Events and return resulting stream.
    rx.map(|msg| match msg {
        Message::UserId(my_id) => Ok(Event::default().event("user").data(my_id.to_string())),
        Message::Reply(reply) => Ok(Event::default().data(reply)),
    })
}

pub fn retry(self, duration: Duration) -> Event

Set Server-sent event retry Retry timeout field (“retry:”)

pub fn id<T: Into<String>>(self, id: T) -> Event

Set Server-sent event id Identifier field (“id:”)