Module warp::filters::ws

Websockets Filters

Structs

Message
A WebSocket message.

WebSocket
A websocket Stream and Sink, provided to ws filters.

Ws
Extracted by the ws filter, and used to finish an upgrade.

Functions

ws
Creates a Websocket Filter.