Trait warp::redirect::AsLocation

pub trait AsLocation: Sealed { }

Trait for redirect locations. Currently only a Uri can be used in redirect. This sealed trait exists to allow adding possibly new impls so other arguments could be accepted, like maybe just warp::redirect("/v2").