Module warp::filters::body

Body filters

Filters that extract a body for a route.

Structs

BodyDeserializeError
An error used in rejections when deserializing a request body fails.

Functions

aggregate
Returns a Filter that matches any request and extracts a Future of an aggregated body.

bytes
Returns a Filter that matches any request and extracts a Future of a concatenated body.

content_length_limit
Require a content-length header to have a value no greater than some limit.

form
Returns a Filter that matches any request and extracts a Future of a form encoded body.

json
Returns a Filter that matches any request and extracts a Future of a JSON-decoded body.

stream
Create a Filter that extracts the request body as a futures::Stream.