warp::filters

Module fs

File System Filters

Structs

ArcPath ğŸ”’

BadRange ğŸ”’

Conditionals ğŸ”’

File

A file response.

FileOpenError ğŸ”’

FilePermissionError ğŸ”’

Enums

Cond ğŸ”’

Constants

DEFAULT_READ_BUF_SIZE ğŸ”’

Functions

bytes_range ğŸ”’

conditionals ğŸ”’

dir

Creates a Filter that serves a directory at the base path joined by the request path.

file

Creates a Filter that serves a File at the path.

file_conditional ğŸ”’

file_metadata ğŸ”’

file_reply ğŸ”’

file_stream ğŸ”’

get_block_size ğŸ”’

optimal_buf_size ğŸ”’

path_from_tail ğŸ”’

reserve_at_least ğŸ”’

sanitize_path ğŸ”’