Module warp::filters::fs

File System Filters

Structs

ArcPath 🔒

BadRange 🔒

Conditionals 🔒

File

A file response.

FileOpenError 🔒

FilePermissionError 🔒

Enums

Cond 🔒

Constants

DEFAULT_READ_BUF_SIZE 🔒

Functions

bytes_range 🔒

conditionals 🔒

dir

Creates a Filter that serves a directory at the base path joined by the request path.

file

Creates a Filter that serves a File at the path.

file_conditional 🔒

file_metadata 🔒

file_reply 🔒

file_stream 🔒

get_block_size 🔒

optimal_buf_size 🔒

path_from_tail 🔒

reserve_at_least 🔒

sanitize_path 🔒