warp::filters

Module header

Header Filters

These filters are used to interact with the Request HTTP headers. Some of them, like exact and exact_ignore_case, are just predicates, they don’t extract any values. The header filter allows parsing a type from any header.

Functions

exact

Create a Filter that requires a header to match the value exactly.

exact_ignore_case

Create a Filter that requires a header to match the value exactly.

header

Create a Filter that tries to parse the specified header.

headers_cloned

Create a Filter that returns a clone of the request’s HeaderMap.

optional

Create a Filter that tries to parse the specified header, if it exists.

value

Create a Filter that gets a HeaderValue for the name.