Module warp::filters::method

HTTP Method filters.

The filters deal with the HTTP Method part of a request. Several here will match the request Method, and if not matched, will reject the request with a 405 Method Not Allowed.

There is also warp::method(), which never rejects a request, and just extracts the method to be used in your filter chains.

Re-exports

pub use self::v2::delete as delete2;

pub use self::v2::get as get2;

pub use self::v2::head;

pub use self::v2::options;

pub use self::v2::patch;

pub use self::v2::post as post2;

pub use self::v2::put as put2;

Modules

v2

HTTP Method Filters

Functions

method

Extract the Method from the request.