Module warp::filters::method::v2

HTTP Method Filters

These filters deal with the HTTP Method part of a request. They match the request Method, and if not matched, will reject the request with a 405 Method Not Allowed.

Functions

delete
Create a Filter that requires the request method to be DELETE.

get
Create a Filter that requires the request method to be GET.

head
Create a Filter that requires the request method to be HEAD.

options
Create a Filter that requires the request method to be OPTIONS.

patch
Create a Filter that requires the request method to be PATCH.

post
Create a Filter that requires the request method to be POST.

put
Create a Filter that requires the request method to be PUT.