Function warp::filters::query::query

pub fn query<T: DeserializeOwned + Send + 'static>(
) -> impl Filter<Extract = (T,), Error = Rejection> + Copy

Creates a Filter that decodes query parameters to the type T.

If cannot decode into a T, the request is rejected with a 400 Bad Request.