Module warp::filters::trace

tracing filters.

tracing is a framework for instrumenting Rust programs to collect scoped, structured, and async-aware diagnostics. This module provides a set of filters for instrumenting Warp applications with tracing spans. Spans can be used to associate individual events with a request, and track contexts through the application.

Structs

Info

Information about the request/response that can be used to prepare log lines.

Trace

Decorates a Filter to create a tracing span for requests and responses.

Functions

named

Create a wrapping filter that instruments every request with a tracing Span at the DEBUG level representing a named context.

request

Create a wrapping filter that instruments every request with a tracing Span at the INFO level, containing a summary of the request. Additionally, if the DEBUG level is enabled, the span will contain an event recording the request’s headers.

trace

Create a wrapping filter that instruments every request with a custom tracing Span provided by a function.