Struct warp::filters::ws::Message

pub struct Message { /* private fields */ }

A WebSocket message.

This will likely become a non-exhaustive enum in the future, once that language feature has stabilized.

Implementations

impl Message

pub fn text<S: Into<String>>(s: S) -> Message

Construct a new Text Message.

pub fn binary<V: Into<Vec<u8>>>(v: V) -> Message

Construct a new Binary Message.

pub fn ping<V: Into<Vec<u8>>>(v: V) -> Message

Construct a new Ping Message.

pub fn pong<V: Into<Vec<u8>>>(v: V) -> Message

Construct a new Pong Message.

Note that one rarely needs to manually construct a Pong message because the underlying tungstenite socket automatically responds to the Ping messages it receives. Manual construction might still be useful in some cases like in tests or to send unidirectional heartbeats.

pub fn close() -> Message

Construct the default Close Message.

pub fn close_with(
code: impl Into<u16>,
reason: impl Into<Cow<'static, str>>
) -> Message

Construct a Close Message with a code and reason.

pub fn is_text(&self) -> bool

Returns true if this message is a Text message.

pub fn is_binary(&self) -> bool

Returns true if this message is a Binary message.

pub fn is_close(&self) -> bool

Returns true if this message a is a Close message.

pub fn is_ping(&self) -> bool

Returns true if this message is a Ping message.

pub fn is_pong(&self) -> bool

Returns true if this message is a Pong message.

pub fn close_frame(&self) -> Option<(u16, &str)>

Try to get the close frame (close code and reason)

pub fn to_str(&self) -> Result<&str, ()>

Try to get a reference to the string text, if this is a Text message.

pub fn as_bytes(&self) -> &[u8]

Return the bytes of this message, if the message can contain data.

pub fn into_bytes(self) -> Vec<u8>

Destructure this message into binary data.