Struct warp::filters::ws::Ws

pub struct Ws { /* private fields */ }

Extracted by the ws filter, and used to finish an upgrade.

Implementations

impl Ws

pub fn on_upgrade<F, U>(self, func: F) -> impl Replywhere
F: FnOnce(WebSocket) -> U + Send + 'static,
U: Future<Output = ()> + Send + 'static,

Finish the upgrade, passing a function to handle the WebSocket.

The passed function must return a Future.

pub fn max_send_queue(self, max: usize) -> Self

Set the size of the internal message send queue.

pub fn max_message_size(self, max: usize) -> Self

Set the maximum message size (defaults to 64 megabytes)

pub fn max_frame_size(self, max: usize) -> Self

Set the maximum frame size (defaults to 16 megabytes)