Module warp::reject

Rejections

Part of the power of the Filter system is being able to reject a request from a filter chain. This allows for filters to be combined with or, so that if one side of the chain finds that a request doesn’t fulfill its requirements, the other side can try to process the request.

Many of the built-in filters will automatically reject the request with an appropriate rejection. However, you can also build new custom Filters and still want other routes to be matchable in the case a predicate doesn’t hold.

As a request is processed by a Filter chain, the rejections are accumulated into a list contained by the Rejection type. Rejections from filters can be handled using Filter::recover. This is a convenient way to map rejections into a Reply.

For a more complete example see the Rejection Example from the repository.

Example

use warp::{reply, Reply, Filter, reject, Rejection, http::StatusCode};

#[derive(Debug)]
struct InvalidParameter;

impl reject::Reject for InvalidParameter {};

// Custom rejection handler that maps rejections into responses.
async fn handle_rejection(err: Rejection) -> Result<impl Reply, std::convert::Infallible> {
    if err.is_not_found() {
        Ok(reply::with_status("NOT_FOUND", StatusCode::NOT_FOUND))
    } else if let Some(e) = err.find::<InvalidParameter>() {
        Ok(reply::with_status("BAD_REQUEST", StatusCode::BAD_REQUEST))
    } else {
        eprintln!("unhandled rejection: {:?}", err);
        Ok(reply::with_status("INTERNAL_SERVER_ERROR", StatusCode::INTERNAL_SERVER_ERROR))
    }
}

#[tokio::main]
async fn main() {

    // Filter on `/:id`, but reject with InvalidParameter if the `id` is `0`.
    // Recover from this rejection using a custom rejection handler.
    let route = warp::path::param()
        .and_then(|id: u32| async move {
            if id == 0 {
                Err(warp::reject::custom(InvalidParameter))
            } else {
                Ok("id is valid")
            }
        })
        .recover(handle_rejection);

    warp::serve(route).run(([127, 0, 0, 1], 3030)).await;
}

Structs

InvalidHeader

Invalid request header

InvalidQuery

Invalid query

LengthRequired

A content-length header is required

MethodNotAllowed

HTTP method not allowed

MissingCookie

Missing cookie

MissingHeader

Missing request header

PayloadTooLarge

The request payload is too large

Rejection

Rejection of a request by a Filter.

UnsupportedMediaType

The request’s content-type is not supported

Traits

Reject

A marker trait to ensure proper types are used for custom rejections.

Functions

custom

Rejects a request with a custom cause.

not_found

Rejects a request with 404 Not Found.

reject

Rejects a request with 404 Not Found.