Function warp::reject::custom

pub fn custom<T: Reject>(err: T) -> Rejection

Rejects a request with a custom cause.

A recover filter should convert this Rejection into a Reply, or else this will be returned as a 500 Internal Server Error.