Trait warp::reject::sealed::CombineRejection

pub trait CombineRejection<E>: Send + Sized {
    type One: IsReject + From<Self> + From<E> + Into<Rejection>;
    type Combined: IsReject;
    fn combine(self, other: E) -> Self::Combined;
}

Associated Types

type One: IsReject + From<Self> + From<E> + Into<Rejection>

The type that should be returned when only 1 of the two “rejections” occurs.

For example:

warp::any().and(warp::path("foo")) has the following steps:

Since this is and, only one of the rejections will occur, and as soon as it does, it will be returned.

warp::any() rejects with Never. So, it will never return Never.

warp::path() rejects with Rejection. It may return Rejection.

Thus, if the above filter rejects, it will definitely be Rejection.

type Combined: IsReject

The type that should be returned when both rejections occur, and need to be combined.

Required methods

fn combine(self, other: E) -> Self::Combined