Struct warp::reject::Rejection

pub struct Rejection {
    reason: Reason,
}

Rejection of a request by a Filter.

See the reject documentation for more.

Fields
reason: Reason
Implementations

impl Rejection

fn known(known: Known) -> Self

fn custom(other: Box<dyn Cause>) -> Self

pub fn find<T: 'static>(&self) -> Option<&T>

Searches this Rejection for a specific cause.

A Rejection will accumulate causes over a Filter chain. This method can search through them and return the first cause of this type.

Example

#[derive(Debug)]
struct Nope;

impl warp::reject::Reject for Nope {}

let reject = warp::reject::custom(Nope);

if let Some(nope) = reject.find::<Nope>() {
   println!("found it: {:?}", nope);
}

pub fn is_not_found(&self) -> bool

Returns true if this Rejection was made via warp::reject::not_found.

Example

let rejection = warp::reject();

assert!(rejection.is_not_found());